Move `generatePayload` back to feature service file instead of util
